requirements.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
mysqlclient==2.1.1
protobuf==3.20.3
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
